requirements.txt
python-telegram-bot
supabase
requests
